move generic helper stuff to generic helper location, also fix an unforseen edge case with the new chain placement
